+++ Nethereum.BlockProcessing/Processors/PostProcessors/BlockPostProcessor.cs
@@ -4,17 +4,19 @@ using Nethereum.Contracts.Services;
 using Nethereum.RPC.Eth.Services;
 using Nethereum.Web3;
 using System;
+using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Nethereum.BlockchainProcessing.Processors.PostProcessors
 {
     public class BlockPostProcessor : BlockProcessor
     {
-        public BlockPostProcessor(IWeb3 web3, IBlockHandler blockHandler, ITransactionProcessor transactionProcessor) : base(web3, blockHandler, transactionProcessor)
+        public BlockPostProcessor(IWeb3 web3, IBlockHandler blockHandler, ITransactionProcessor transactionProcessor) 
+            : base(web3, blockHandler, transactionProcessor)
         {
         }
 
-        public override async Task ProcessBlockAsync(ulong blockNumber)
+        public override async Task ProcessBlockAsync(BigInteger blockNumber)
         {
             var block = await BlockProxy.Blocks
                 .GetBlockWithTransactionsByNumber.SendRequestAsync(blockNumber.ToHexBigInteger())
