+++ Nethereum.BlockchainProcessing/Processing/IBlockProgressRepository.cs
@@ -1,11 +1,12 @@
 ﻿using System;
+using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Nethereum.BlockchainProcessing.Processing
 {
     public interface IBlockProgressRepository
     {
-        Task UpsertProgressAsync(ulong blockNumber);
-        Task<ulong?> GetLastBlockNumberProcessedAsync();
+        Task UpsertProgressAsync(BigInteger blockNumber);
+        Task<BigInteger?> GetLastBlockNumberProcessedAsync();
     }
 }
