+++ Nethereum.LogProcessing/ILogsProcessor.cs
@@ -1,5 +1,7 @@
 ﻿using Nethereum.BlockchainProcessing.Processing;
+using Nethereum.Contracts;
 using System;
+using System.Numerics;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -10,7 +12,7 @@ namespace Nethereum.LogProcessing
     {
         event EventHandler OnDisposing;
 
-        Task<ulong> ProcessContinuallyAsync(CancellationToken cancellationToken, Action<LogBatchProcessedArgs> batchCompleteCallback = null);
+        Task<BigInteger> ProcessContinuallyAsync(CancellationToken cancellationToken, Action<LogBatchProcessedArgs> batchCompleteCallback = null);
         Task ProcessContinuallyInBackgroundAsync(CancellationToken cancellationToken, Action<LogBatchProcessedArgs> batchCompleteCallback = null, Action<Exception> fatalErrorCallback = null);
         Task<BlockRange?> ProcessOnceAsync(CancellationToken cancellationToken = default);
     }
